Don't run Rust calls twice on error
Due to how C macros work, `CHECK_FFI`'s input was not the evaluated
result of the arguments, but the literal code itself. That means
that in referring to the argument twice we would actually run the
code twice. In some cases this could do very bad things like actually
triggering a segfault due to use of variables that had been freed.